Change integer variables from int32 to int64, so they match the backend
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/9322
Compute-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>

+++ cli/src/admiral/cmd/commons.go
@@ -205,8 +205,8 @@ var (
 	memoryLimitStr     string
 	memoryLimitDesc    = "Memory limit. Default unit: kb. Units supported: kb/mb/gb. Example: 1024mb"
 	priorityInt        int32
-	maxNumberInstances int32
-	cpuSharesInt       int32
+	maxNumberInstances int64
+	cpuSharesInt       int64
 
 	outputFormat     string
 	outputFormatDesc = "Output format: json, table."
